Fix issues about presets
