+++ common/src/main/java/icu/nullptr/hidemyapplist/common/RiskyPackageUtils.kt
@@ -1,7 +1,7 @@
 package icu.nullptr.hidemyapplist.common
 
 import android.content.pm.ApplicationInfo
-import java.util.zip.ZipFile
+import icu.nullptr.hidemyapplist.common.Utils.checkSplitPackages
 
 object RiskyPackageUtils {
     private const val GMS_PROP = "\u0000c\u0000o\u0000m\u0000.\u0000g\u0000o\u0000o\u0000g\u0000l\u0000e\u0000.\u0000a\u0000n\u0000d\u0000r\u0000o\u0000i\u0000d\u0000.\u0000g\u0000m\u0000s\u0000."
@@ -11,20 +11,21 @@ object RiskyPackageUtils {
 
     fun appHasGMSConnection(query: String) = query in ignoredForRiskyPackagesList
 
-    internal fun tryToAddIntoGMSConnectionList(appInfo: ApplicationInfo, packageName: String, loggerFunction: ((String) -> Unit)?) {
-        if (packageName in ignoredForRiskyPackagesList) return
+    internal fun tryToAddIntoGMSConnectionList(appInfo: ApplicationInfo, packageName: String, loggerFunction: ((String) -> Unit)?): Boolean {
+        if (packageName in ignoredForRiskyPackagesList) return false
 
-        try {
-            ZipFile(appInfo.sourceDir).use { zipFile ->
-                val manifestStr = AppPresets.instance.readManifest(appInfo.sourceDir, zipFile)
+        return checkSplitPackages(appInfo) { key, zipFile ->
+            val manifestStr = AppPresets.instance.readManifest(key, zipFile)
 
-                // Checking with binary because the Android system sucks
-                if (manifestStr.contains(GMS_PROP) || manifestStr.contains(FIREBASE_PROP)) {
-                    if (ignoredForRiskyPackagesList.add(packageName)) {
-                        loggerFunction?.invoke("@appHasGMSConnection $packageName added in ignored packages list")
-                    }
+            // Checking with binary because the Android system sucks
+            if (manifestStr.contains(GMS_PROP) || manifestStr.contains(FIREBASE_PROP)) {
+                if (ignoredForRiskyPackagesList.add(packageName)) {
+                    loggerFunction?.invoke("@appHasGMSConnection $packageName added in ignored packages list")
+                    return@checkSplitPackages true
                 }
             }
-        } catch (_: Throwable) { }
+
+            return@checkSplitPackages false
+        }
     }
 }
