+++ common/src/main/java/icu/nullptr/hidemyapplist/common/Utils.kt
@@ -7,6 +7,7 @@ import android.content.pm.ResolveInfo
 import android.os.Binder
 import android.os.Build
 import java.util.Random
+import java.util.zip.ZipFile
 
 object Utils {
 
@@ -91,5 +92,21 @@ object Utils {
             resolveInfo.activityInfo?.packageName ?:
             resolveInfo.serviceInfo?.packageName ?:
             resolveInfo.providerInfo!!.packageName
+    }
+
+    fun checkSplitPackages(appInfo: ApplicationInfo, onZipFile: (String, ZipFile) -> Boolean): Boolean {
+        val allLocations = setOf(appInfo.sourceDir, appInfo.publicSourceDir) /*+
+                (appInfo.splitSourceDirs ?: arrayOf()) +
+                (appInfo.splitPublicSourceDirs ?: arrayOf())*/
+
+        return allLocations.any { filePath ->
+            ZipFile(filePath).use { zipFile ->
+                if (onZipFile(filePath, zipFile)) {
+                    return true
+                }
+            }
+
+            return false
+        }
     }
 }
