+++ common/src/main/java/icu/nullptr/hidemyapplist/common/app_presets/RootAppsPreset.kt
@@ -3,6 +3,7 @@ package icu.nullptr.hidemyapplist.common.app_presets
 import android.content.pm.ApplicationInfo
 import icu.nullptr.hidemyapplist.common.AppPresets
 import icu.nullptr.hidemyapplist.common.Utils
+import icu.nullptr.hidemyapplist.common.Utils.checkSplitPackages
 
 class RootAppsPreset(private val appPresets: AppPresets) : BasePreset(NAME) {
     companion object {
