+++ common/src/main/java/icu/nullptr/hidemyapplist/common/app_presets/XposedModulesPreset.kt
@@ -1,7 +1,7 @@
 package icu.nullptr.hidemyapplist.common.app_presets
 
 import android.content.pm.ApplicationInfo
-import java.util.zip.ZipFile
+import icu.nullptr.hidemyapplist.common.Utils.checkSplitPackages
 
 class XposedModulesPreset : BasePreset(NAME) {
     companion object {
@@ -11,7 +11,7 @@ class XposedModulesPreset : BasePreset(NAME) {
     override val exactPackageNames = setOf<String>()
 
     override fun canBeAddedIntoPreset(appInfo: ApplicationInfo): Boolean {
-        checkSplitPackages(appInfo) { _, zipFile ->
+        return checkSplitPackages(appInfo) { _, zipFile ->
             // Legacy Xposed method
             if (zipFile.getEntry("assets/xposed_init") != null) {
                 return@checkSplitPackages true
@@ -24,7 +24,5 @@ class XposedModulesPreset : BasePreset(NAME) {
 
             return@checkSplitPackages false
         }
-
-        return false
     }
 }
