+++ xposed/src/main/java/icu/nullptr/hidemyapplist/xposed/HMAService.kt
@@ -36,7 +36,6 @@ import rikka.hidden.compat.ActivityManagerApis
 import java.io.File
 import java.util.concurrent.ExecutorService
 import java.util.concurrent.Executors
-import kotlin.concurrent.thread
 
 class HMAService(val pms: IPackageManager, val pmn: Any?) : IHMAService.Stub() {
